+++ connect/src/main/java/com/flip/connect/domain/model/BaseResponse.java
@@ -14,27 +14,18 @@ public class BaseResponse {
     private Boolean success = false;
     private List<OperationReport> operationReport;
 
-    public Boolean getSuccess() {
+    public Boolean hasSuccess() {
         return success;
-    }
-
-    public void setSuccess(Boolean success) {
-        this.success = success;
     }
 
     public List<OperationReport> getOperationReport() {
         return operationReport;
     }
 
-    public void setOperationReport(List<OperationReport> operationReport) {
-        this.operationReport = operationReport;
-    }
-
     @Override
     public String toString() {
         String aux = "";
         for(OperationReport or : operationReport){
-            Log.e("RequestError", "Success: "+success+", field: "+or.getField()+", message: "+or.getMessage());
             aux+=or.getMessage();
         }
         return aux;
