+++ connect/src/main/java/com/flip/connect/domain/model/user/UserResponse.java
@@ -1,10 +1,7 @@
 package com.flip.connect.domain.model.user;
 
 import com.flip.connect.domain.model.BaseResponse;
-import com.flip.connect.domain.model.auth.OperationReport;
 import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
 
 public class UserResponse extends BaseResponse {
 
@@ -24,7 +21,7 @@ public class UserResponse extends BaseResponse {
         return
                 "UserResponse{" +
                         "operationReport = '" + getOperationReport() + '\'' +
-                        ",success = '" + getSuccess() + '\'' +
+                        ",success = '" + hasSuccess() + '\'' +
                         ",user = '" + user + '\'' +
                         "}";
     }
