+++ connect/src/main/java/com/flip/connect/presentation/views/login/LoginContract.java
@@ -1,6 +1,6 @@
 package com.flip.connect.presentation.views.login;
 
-import com.flip.connect.data.model.TempProfile;
+import com.flip.connect.data.model.tempProfile.TempProfile;
 import com.flip.connect.domain.model.auth.OauthToken;
 import com.flip.connect.presentation.base.BasePresenterContract;
 import com.flip.connect.presentation.base.BaseViewContract;
